Keyboard: Un-parameterize layout files
This configuration was for a long time part of my ansible system
configuration playbooks
(https://github.com/jan-warchol/ansible-system-setup), and it was
parameterized. However, I came to the conclusion that it doesn't make
sense to mix keyboard layout and ansible code, and I'm splitting this
back to its own repo (it actually used to be like that in the beginning,
before I made a subtree merge into ansible-system-setup).

The previous commit in this repo corresponds to
6269389202507770427a6dde9dac374373b47290 in ansible deployment.